login progress
not entirely finished but I got most of the basics down for the login. I have not seen the webpage you guys have to test on the main branch or your branches so I improvised. If you guys can test it and provide feedback when you can that would be great. It most likely doesn't work since it still needs fixing but I am on it right now after this (I am also trying to set up testing which has not been working out).